third_party/blink/web_tests/fast/encoding/meta-in-script-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
PASS: windows-1255


This test passes if the charset is parsed from the meta tag outside the script.
